in class

+++ code/ComplexPlane.cpp
@@ -18,16 +18,16 @@ ComplexPlane::ComplexPlane(float aspectRatio)
 void ComplexPlane::zoomIn()
 {
 	m_zoomCount++;
-	float x_size = BASE_WIDTH * pow(BASE_ZOOM, m_zoomCount);
-	float y_size = BASE_HEIGHT * m_aspectRatio * pow(BASE_ZOOM, m_zoomCount);
+	double x_size = BASE_WIDTH * pow(BASE_ZOOM, m_zoomCount);
+	double y_size = BASE_HEIGHT * m_aspectRatio * pow(BASE_ZOOM, m_zoomCount);
 	m_view.setSize(x_size, y_size);
 };
 
 void ComplexPlane::zoomOut()
 {
 	m_zoomCount--;
-	float x_size = BASE_WIDTH * pow(BASE_ZOOM, m_zoomCount);
-	float y_size = BASE_HEIGHT * m_aspectRatio * pow(BASE_ZOOM, m_zoomCount);
+	double x_size = BASE_WIDTH * pow(BASE_ZOOM, m_zoomCount);
+	double y_size = BASE_HEIGHT * m_aspectRatio * pow(BASE_ZOOM, m_zoomCount);
 	m_view.setSize(x_size, y_size);
 };
